Fix sample code based on improved usage of Image.Translate(). (#2)
The imaging code of the .NET and Java libraries has been updated significantly and now Image.Translate() takes coordinates as one would expect on the page.
Whereas previously the values were based on internally applying the matrix multiplication backwards.  Because of this change the sample needs to be updated to produce the correct output.