Change how we get packages off the raid.
We have frequent intermittent failures where the Nuget packages are corrupted in some fashion. What we currently do is for each sample build and point our source of Nuget
packages at the location on the raid.

I think this is part of the problem, when there's an intermittent hiccup or conflict with what's on the raid.

Instead, let's first copy the packages of interest locally. Then we'll use that as our source to point at.

Also, refactor to remove some redundancy.

This should lead to some performance improvement because now we no longer are dependent on constantly asking dotnet to look at the raid for each sample's build, we just
look at the raid once to get what we need and then are local again.